Publicly publishing the A1MGBot that I created a while back.
My server has been using this to get data about players using the repository DennisV2809 created. All credit for the original idea goes to him. I only took it, expanded it a bit, polished it, and wrapped it in a Discord bot.

If you plan to use this code to create your own Discord bots, wouldn't hurt to give me credit, but I don't really care if you don't.